+++ src/handlers/profile/lookup.rs
@@ -1,5 +1,9 @@
 use {
-    super::super::HANDLER_TASK_METRICS,
+    super::{
+        super::HANDLER_TASK_METRICS,
+        utils::{is_name_format_correct, is_name_in_allowed_zones, is_name_length_correct},
+        ALLOWED_ZONES,
+    },
     crate::{database::helpers::get_name_and_addresses_by_name, error::RpcError, state::AppState},
     axum::{
         extract::{Path, State},
@@ -27,14 +31,25 @@ async fn handler_internal(
     state: State<Arc<AppState>>,
     Path(name): Path<String>,
 ) -> Result<Response, RpcError> {
-    match get_name_and_addresses_by_name(name, &state.postgres).await {
+    // Check if the name is in the correct format
+    if !is_name_format_correct(&name) {
+        return Err(RpcError::InvalidNameFormat(name));
+    }
+
+    // Check if the name length is correct
+    if !is_name_length_correct(&name) {
+        return Err(RpcError::InvalidNameLength(name));
+    }
+
+    // Check is name in the allowed zones
+    if !is_name_in_allowed_zones(&name, &ALLOWED_ZONES) {
+        return Err(RpcError::InvalidNameZone(name));
+    }
+
+    match get_name_and_addresses_by_name(name.clone(), &state.postgres).await {
         Ok(response) => Ok(Json(response).into_response()),
         Err(e) => match e {
-            SqlxError::RowNotFound => {
-                // Respond with a "Not Found" status code if name was found
-                let not_found_response = (StatusCode::NOT_FOUND, "Name is not registered");
-                Ok(not_found_response.into_response())
-            }
+            SqlxError::RowNotFound => return Err(RpcError::NameNotFound(name)),
             _ => {
                 // Handle other types of errors
                 error!("Failed to lookup name: {}", e);